Converting double folds to int

+++ test/lifepro_flexstride_pro_test.dart
@@ -103,7 +103,7 @@ void main() {
         ),
       ),
     ]) {
-      final sum = testPair.data.fold<double>(0.0, (a, b) => a + b);
+      final sum = testPair.data.fold<int>(0, (a, b) => a + b);
       test("$sum ${testPair.data.length}", () async {
         final xTrainer = DeviceFactory.getGenericFTMSCrossTrainer();
         xTrainer.initFlag();
